Refactor to useModal hook
